fixup! fix(ngcc): do not attempt compilation when analysis fails
